+++ iblrig/commands.py
@@ -365,3 +365,22 @@ def flush():
     bpod = Bpod(hardware_settings['device_bpod']['COM_BPOD'])
     bpod.flush()
     bpod.close()
+
+
+def remove_bonsai_layouts():
+    """Delete all BONSAI .layout files - if they are backed up with a .layout_template file."""
+    from iblrig.constants import BASE_PATH
+
+    layout_files = [x for x in BASE_PATH.glob('**/*.bonsai.layout') if x.with_suffix('.layout_template').exists()]
+    if len(layout_files) == 0:
+        print('No layout files found.')
+        return
+    print('The following files will be deleted:')
+    for f in layout_files:
+        print(f'- {f.name}')
+    if input('\nContinue? [Y/n] ').lower() in ('y', ''):
+        for f in layout_files:
+            f.unlink()
+        print(f'{len(layout_files)} file{"s" if len(layout_files) > 1 else ""} deleted.')
+    else:
+        print('No files deleted.')