+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Changelog
 =========
 
+8.27.4
+------
+* added: `_iblrig_tasks_spontaneousBpod` - a spontaneous task that includes Bpod spacers
+
+8.27.3
+------
+* changed: reset camera(s) prior to starting task when inconsistencies have been detected
+* changed: include stock subjects if user is stock manager
+* changed: skip validation of Ambient Module if device_bpod.USE_AMBIENT_MODULE is false
+* added: `remove_bonsai_layouts` command for troubleshooting BONSAI GUIs not appearing 
+
+8.27.2
+------
+* fixed: error when using task arguments of type bool
+
 8.27.1
 ------
 * fixed: resolved warning about chained assignment in TrainingChoiceWorld.compute_performance()